requirements.txt
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
sacremoses>=0.1.1
frida-tools>=12.0.0
frida>=16.0.0
psutil>=5.9.0
requests>=2.31.0
chardet>=5.0.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
UnityPy>=1.20.0
Pillow>=10.0.0
